+++ routes/apis.js
@@ -11,15 +11,47 @@
 var app = module.parent.exports.app,
   config = module.parent.exports.config,
   sequelize = module.parent.exports.sequelize,
+  epilogue = module.parent.exports.epilogue,
   logger = module.parent.exports.logger;
-  // ## Models
-  /*
-  var Client  = require('../models/clients.js'),
+// ## Models
+var Client  = require('../models/clients.js'),
   Provider  = require('../models/providers.js');
   /* models:end */
 
 // ## Public Rest
 // --------------------------------------
+
+exports.sequelize = sequelize;
+
+// Initialize epilogue
+epilogue.initialize({
+  app       : app,
+  sequelize : sequelize
+});
+
+// Create REST resources
+var clientResource = epilogue.resource({
+  model: Client,
+  endpoints: ['/api/v1/clients', '/api/v1/clients/:id']
+});
+
+var providerResource = epilogue.resource({
+  model: Provider,
+  endpoints: ['/api/v1/providers', '/api/v1/providers/:idProvider']
+});
+
+sequelize.sync({force:true});
+
+/*
+app.get('/clients', function(req, res){
+  Client.findAll().then(function(result){
+    res.json(result);
+  });
+});
+*/
+
+
+
 
 // CORS Interceptors
 if (config.cors && config.cors === "enabled") {
